clean app.jsx
