+++ src/pages/Projects/Projects.jsx
@@ -98,61 +98,162 @@ const Projects = () => {
     };
 
     return (
-        <div className="projects-main-box">
-            <div className={`projects-category ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
-                <ul className='tab-list'>
-                    <li
-                        style={{
-                            color: pokeTheme_TWO,
-                            cursor: 'pointer',
-                            borderBottom: selectedTab === 'All' ? `2px solid ${pokeTheme_TWO}` : 'none',
-                            fontWeight: selectedTab === 'All' ? `700` : '',
-                        }}
-                        onClick={() => handleTabClick('All')}
-                    >
-                        All
-                    </li>
-                    <li
-                        style={{
-                            color: pokeTheme_TWO,
-                            cursor: 'pointer',
-                            borderBottom: selectedTab === 'Coding' ? `2px solid ${pokeTheme_TWO}` : 'none',
-                            fontWeight: selectedTab === 'Coding' ? `700` : '',
-                        }}
-                        onClick={() => handleTabClick('Coding')}
-                    >
-                        Coding
-                    </li>
-                    <li
-                        style={{
-                            color: pokeTheme_TWO,
-                            cursor: 'pointer',
-                            borderBottom: selectedTab === 'Research' ? `2px solid ${pokeTheme_TWO}` : 'none',
-                            fontWeight: selectedTab === 'Research' ? `700` : '',
-                        }}
-                        onClick={() => handleTabClick('Research')}
-                    >
-                        Research
-                    </li>
-                    <li
-                        style={{
-                            color: pokeTheme_TWO,
-                            cursor: 'pointer',
-                            borderBottom: selectedTab === 'Art' ? `2px solid ${pokeTheme_TWO}` : 'none',
-                            fontWeight: selectedTab === 'Art' ? `700` : '',
-                        }}
-                        onClick={() => handleTabClick('Art')}
-                    >
-                        Art
-                    </li>
-                </ul>
-            </div>
+        <div className="content-item projects1">
+            <div className="projects-main-box">
+                <div className={`projects-category ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
+                    <ul className='tab-list'>
+                        <li
+                            style={{
+                                color: pokeTheme_TWO,
+                                cursor: 'pointer',
+                                borderBottom: selectedTab === 'All' ? `2px solid ${pokeTheme_TWO}` : 'none',
+                                fontWeight: selectedTab === 'All' ? `700` : '',
+                            }}
+                            onClick={() => handleTabClick('All')}
+                        >
+                            All
+                        </li>
+                        <li
+                            style={{
+                                color: pokeTheme_TWO,
+                                cursor: 'pointer',
+                                borderBottom: selectedTab === 'Coding' ? `2px solid ${pokeTheme_TWO}` : 'none',
+                                fontWeight: selectedTab === 'Coding' ? `700` : '',
+                            }}
+                            onClick={() => handleTabClick('Coding')}
+                        >
+                            Coding
+                        </li>
+                        <li
+                            style={{
+                                color: pokeTheme_TWO,
+                                cursor: 'pointer',
+                                borderBottom: selectedTab === 'Research' ? `2px solid ${pokeTheme_TWO}` : 'none',
+                                fontWeight: selectedTab === 'Research' ? `700` : '',
+                            }}
+                            onClick={() => handleTabClick('Research')}
+                        >
+                            Research
+                        </li>
+                        <li
+                            style={{
+                                color: pokeTheme_TWO,
+                                cursor: 'pointer',
+                                borderBottom: selectedTab === 'Art' ? `2px solid ${pokeTheme_TWO}` : 'none',
+                                fontWeight: selectedTab === 'Art' ? `700` : '',
+                            }}
+                            onClick={() => handleTabClick('Art')}
+                        >
+                            Art
+                        </li>
+                    </ul>
+                </div>
 
 
-            <div className={`projects-list ${displayType ? 'light' : 'dark'} ${isTransitioning ? 'hidden' : ''} ${selectedTab === 'All' ? 'all' : 'single'}`}>
-                {selectedTab === 'All' && (
-                    <>
-                        <div className='group-layout'>
+                <div className={`projects-list ${displayType ? 'light' : 'dark'} ${isTransitioning ? 'hidden' : ''} ${selectedTab === 'All' ? 'all' : 'single'}`}>
+                    {selectedTab === 'All' && (
+                        <>
+                            <div className='group-layout'>
+                                <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopCodingInterval} onMouseLeave={startCodingInterval} >
+                                    <div className='tabs'>
+                                        {[0, 1, 2, 3, 4].map(index => (
+                                            <button
+                                                key={index}
+                                                className='tab'
+                                                style={{ borderBottom: codingSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
+                                                onClick={() => handleClick(index)}
+                                            >
+                                                {`${index + 1}`}
+                                            </button>
+                                        ))}
+                                    </div>
+                                    <div className='slide-container'>
+                                        {CodeProjects.map((project, index) => (
+                                            <div key={index} className={`slide-content-opacity ${currentSlideClass(index, codingSlide)}`}>
+                                                <div className='project-info'>
+                                                    <div className='project-title'>
+                                                        <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
+                                                            {project.title}
+                                                        </a>
+                                                    </div>
+                                                    <div className='project-desc'>
+                                                        <p>{project.desc}</p>
+                                                    </div>
+                                                </div>
+                                            </div>
+                                        ))}
+                                    </div>
+                                </div>
+
+                                <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopResearchInterval} onMouseLeave={startResearchInterval} >
+                                    <div className='tabs'>
+                                        {[0, 1].map(index => (
+                                            <button
+                                                key={index}
+                                                className='tab'
+                                                style={{ borderBottom: researchSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
+                                                onClick={() => handleResearchClick(index)}
+                                            >
+                                                {`${index + 1}`}
+                                            </button>
+                                        ))}
+                                    </div>
+                                    <div className='slide-container'>
+                                        {ResearchProjects.map((project, index) => (
+                                            <div key={index} className={`slide-content-opacity ${currentSlideClass(index, researchSlide)}`}>
+                                                <div className='project-info'>
+                                                    <div className='project-title'>
+                                                        <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
+                                                            {project.title}
+                                                        </a>
+                                                    </div>
+                                                    <div className='project-desc'>
+                                                        <p>{project.desc}</p>
+                                                    </div>
+                                                </div>
+                                            </div>
+                                        ))}
+                                    </div>
+                                </div>
+                            </div>
+
+
+                            <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopArtInterval} onMouseLeave={startArtInterval} >
+                                <div className='tabs'>
+                                    {[0, 1].map(index => (
+                                        <button
+                                            key={index}
+                                            className='tab'
+                                            style={{ borderBottom: artSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
+                                            onClick={() => handleArtClick(index)}
+                                        >
+                                            {`${index + 1}`}
+                                        </button>
+                                    ))}
+                                </div>
+                                <div className='slide-container'>
+                                    {ArtProjects.map((project, index) => (
+                                        <div key={index} className={`slide-content-opacity ${currentSlideClass(index, artSlide)}`}>
+                                            <div className='project-info'>
+                                                <div className='project-title'>
+                                                    <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
+                                                        {project.title}
+                                                    </a>
+                                                </div>
+                                                <div className='project-desc art'>
+                                                    <img src={project.imgContent} className='art' />
+                                                </div>
+                                            </div>
+                                        </div>
+                                    ))}
+                                </div>
+                            </div>
+                        </>
+                    )}
+
+                    {selectedTab === 'Coding' && (
+                        <>
+                            <div className='single-select' />
                             <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopCodingInterval} onMouseLeave={startCodingInterval} >
                                 <div className='tabs'>
                                     {[0, 1, 2, 3, 4].map(index => (
@@ -183,7 +284,13 @@ const Projects = () => {
                                     ))}
                                 </div>
                             </div>
+                            <div className='single-select' />
+                        </>
+                    )}
 
+                    {selectedTab === 'Research' && (
+                        <>
+                            <div className='single-select' />
                             <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopResearchInterval} onMouseLeave={startResearchInterval} >
                                 <div className='tabs'>
                                     {[0, 1].map(index => (
@@ -214,152 +321,47 @@ const Projects = () => {
                                     ))}
                                 </div>
                             </div>
-                        </div>
+                            <div className='single-select' />
+                        </>
+                    )}
 
-
-                        <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopArtInterval} onMouseLeave={startArtInterval} >
-                            <div className='tabs'>
-                                {[0, 1].map(index => (
-                                    <button
-                                        key={index}
-                                        className='tab'
-                                        style={{ borderBottom: artSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
-                                        onClick={() => handleArtClick(index)}
-                                    >
-                                        {`${index + 1}`}
-                                    </button>
-                                ))}
-                            </div>
-                            <div className='slide-container'>
-                                {ArtProjects.map((project, index) => (
-                                    <div key={index} className={`slide-content-opacity ${currentSlideClass(index, artSlide)}`}>
-                                        <div className='project-info'>
-                                            <div className='project-title'>
-                                                <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
-                                                    {project.title}
-                                                </a>
-                                            </div>
-                                            <div className='project-desc art'>
-                                                <img src={project.imgContent} className='art' />
+                    {selectedTab === 'Art' && (
+                        <>
+                            <div className='single-select' />
+                            <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopArtInterval} onMouseLeave={startArtInterval} >
+                                <div className='tabs'>
+                                    {[0, 1].map(index => (
+                                        <button
+                                            key={index}
+                                            className='tab'
+                                            style={{ borderBottom: artSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
+                                            onClick={() => handleArtClick(index)}
+                                        >
+                                            {`${index + 1}`}
+                                        </button>
+                                    ))}
+                                </div>
+                                <div className='slide-container'>
+                                    {ArtProjects.map((project, index) => (
+                                        <div key={index} className={`slide-content-opacity ${currentSlideClass(index, artSlide)}`}>
+                                            <div className='project-info art'>
+                                                <div className='project-title'>
+                                                    <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
+                                                        {project.title}
+                                                    </a>
+                                                </div>
+                                                <div className='project-desc art'>
+                                                    <img src={project.imgContent} className='art' />
+                                                </div>
                                             </div>
                                         </div>
-                                    </div>
-                                ))}
+                                    ))}
+                                </div>
                             </div>
-                        </div>
-                    </>
-                )}
-
-                {selectedTab === 'Coding' && (
-                    <>
-                        <div className='single-select' />
-                        <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopCodingInterval} onMouseLeave={startCodingInterval} >
-                            <div className='tabs'>
-                                {[0, 1, 2, 3, 4].map(index => (
-                                    <button
-                                        key={index}
-                                        className='tab'
-                                        style={{ borderBottom: codingSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
-                                        onClick={() => handleClick(index)}
-                                    >
-                                        {`${index + 1}`}
-                                    </button>
-                                ))}
-                            </div>
-                            <div className='slide-container'>
-                                {CodeProjects.map((project, index) => (
-                                    <div key={index} className={`slide-content-opacity ${currentSlideClass(index, codingSlide)}`}>
-                                        <div className='project-info'>
-                                            <div className='project-title'>
-                                                <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
-                                                    {project.title}
-                                                </a>
-                                            </div>
-                                            <div className='project-desc'>
-                                                <p>{project.desc}</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
-                        </div>
-                        <div className='single-select' />
-                    </>
-                )}
-
-                {selectedTab === 'Research' && (
-                    <>
-                        <div className='single-select' />
-                        <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopResearchInterval} onMouseLeave={startResearchInterval} >
-                            <div className='tabs'>
-                                {[0, 1].map(index => (
-                                    <button
-                                        key={index}
-                                        className='tab'
-                                        style={{ borderBottom: researchSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
-                                        onClick={() => handleResearchClick(index)}
-                                    >
-                                        {`${index + 1}`}
-                                    </button>
-                                ))}
-                            </div>
-                            <div className='slide-container'>
-                                {ResearchProjects.map((project, index) => (
-                                    <div key={index} className={`slide-content-opacity ${currentSlideClass(index, researchSlide)}`}>
-                                        <div className='project-info'>
-                                            <div className='project-title'>
-                                                <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
-                                                    {project.title}
-                                                </a>
-                                            </div>
-                                            <div className='project-desc'>
-                                                <p>{project.desc}</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
-                        </div>
-                        <div className='single-select' />
-                    </>
-                )}
-
-                {selectedTab === 'Art' && (
-                    <>
-                        <div className='single-select' />
-                        <div className={`projects ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }} onMouseEnter={stopArtInterval} onMouseLeave={startArtInterval} >
-                            <div className='tabs'>
-                                {[0, 1].map(index => (
-                                    <button
-                                        key={index}
-                                        className='tab'
-                                        style={{ borderBottom: artSlide === index ? `2px solid ${pokeTheme_TWO}` : 'none' }}
-                                        onClick={() => handleArtClick(index)}
-                                    >
-                                        {`${index + 1}`}
-                                    </button>
-                                ))}
-                            </div>
-                            <div className='slide-container'>
-                                {ArtProjects.map((project, index) => (
-                                    <div key={index} className={`slide-content-opacity ${currentSlideClass(index, artSlide)}`}>
-                                        <div className='project-info art'>
-                                            <div className='project-title'>
-                                                <a href={project.link} style={{ color: pokeTheme_TWO }} target="_blank" rel="noopener noreferrer">
-                                                    {project.title}
-                                                </a>
-                                            </div>
-                                            <div className='project-desc art'>
-                                                <img src={project.imgContent} className='art' />
-                                            </div>
-                                        </div>
-                                    </div>
-                                ))}
-                            </div>
-                        </div>
-                        <div className='single-select' />
-                    </>
-                )}
+                            <div className='single-select' />
+                        </>
+                    )}
+                </div>
             </div>
         </div>
     );
