+++ src/pages/Skills/Skills.jsx
@@ -61,97 +61,99 @@ const Skills = () => {
     };
 
     return (
-        <div className={`skills-main-box ${selectedSkill ? 'visible' : ''}`}>
-            <div className='skill-list' style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
-                <div className={`skill-header ${displayType ? 'light' : 'dark'}`}>
-                    <h4>Current Tech Stack and Resources</h4>
-                </div>
-                <div className={`skill-icons`}>
-                    {skillsChunk1.map(skill => (
-                        <div
-                            key={skill.name}
-                            className={`skill-item ${displayType ? 'light' : 'dark'} `}
-                            style={{ boxShadow: selectedSkill?.name === skill.name ? `0 0 1rem ${pokeTheme_TWO}` : '' }}
-                            onClick={() => handleSkillClick(skill)}
-                        >
-                            <img
-                                src={skill.imgSrc}
-                                alt={skill.name}
-                                title={skill.name}
-                                className="skill-icon"
-                            />
-                            <p className={`skill-name ${displayType ? 'light' : 'dark'}`}>{skill.name}</p>
-                        </div>
-                    ))}
-                </div>
-            </div>
-
-            <div className={`skill-detail ${selectedSkill ? 'visible' : ''}`}>
-                <div className={`text-detail ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
+        <div className="content-item skills">
+            <div className={`skills-main-box ${selectedSkill ? 'visible' : ''}`}>
+                <div className='skill-list' style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
                     <div className={`skill-header ${displayType ? 'light' : 'dark'}`}>
-                        <h4>My Experience</h4>
+                        <h4>Current Tech Stack and Resources</h4>
                     </div>
-                    <div className={`usage-example ${displayType ? 'light' : 'dark'}`}>
-                        {selectedSkill && (
-                            <>
-                                <div className='usage-header'>
-                                    <p style={{ color: pokeTheme_TWO }}>{selectedSkill.name}</p>
-                                    <p style={{ color: pokeTheme_TWO }}>{selectedSkill.yoe}</p>
-                                </div>
-                                <div className='usage-desc'>
-                                    <p>{selectedSkill.desc}</p>
-                                </div>
-                            </>
-                        )}
-                    </div>
-                </div>
-                <div className={`code-detail ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
-                    <div className={`skill-header ${displayType ? 'light' : 'dark'}`}>
-                        <h4 style={{ marginBottom: '0.625rem' }}>Current Mood</h4>
-                    </div>
-                    <div className={`code-example ${displayType ? 'light' : 'dark'}`}>
-                        {selectedSkill && (
-                            <div className='code-usage' style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
-                                <img src={selectedSkill.meme} />
+                    <div className={`skill-icons`}>
+                        {skillsChunk1.map(skill => (
+                            <div
+                                key={skill.name}
+                                className={`skill-item ${displayType ? 'light' : 'dark'} `}
+                                style={{ boxShadow: selectedSkill?.name === skill.name ? `0 0 1rem ${pokeTheme_TWO}` : '' }}
+                                onClick={() => handleSkillClick(skill)}
+                            >
+                                <img
+                                    src={skill.imgSrc}
+                                    alt={skill.name}
+                                    title={skill.name}
+                                    className="skill-icon"
+                                />
+                                <p className={`skill-name ${displayType ? 'light' : 'dark'}`}>{skill.name}</p>
                             </div>
-                        )}
+                        ))}
+                    </div>
+                </div>
+
+                <div className={`skill-detail ${selectedSkill ? 'visible' : ''}`}>
+                    <div className={`text-detail ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
+                        <div className={`skill-header ${displayType ? 'light' : 'dark'}`}>
+                            <h4>My Experience</h4>
+                        </div>
+                        <div className={`usage-example ${displayType ? 'light' : 'dark'}`}>
+                            {selectedSkill && (
+                                <>
+                                    <div className='usage-header'>
+                                        <p style={{ color: pokeTheme_TWO }}>{selectedSkill.name}</p>
+                                        <p style={{ color: pokeTheme_TWO }}>{selectedSkill.yoe}</p>
+                                    </div>
+                                    <div className='usage-desc'>
+                                        <p>{selectedSkill.desc}</p>
+                                    </div>
+                                </>
+                            )}
+                        </div>
+                    </div>
+                    <div className={`code-detail ${displayType ? 'light' : 'dark'}`} style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
+                        <div className={`skill-header ${displayType ? 'light' : 'dark'}`}>
+                            <h4 style={{ marginBottom: '0.625rem' }}>Current Mood</h4>
+                        </div>
+                        <div className={`code-example ${displayType ? 'light' : 'dark'}`}>
+                            {selectedSkill && (
+                                <div className='code-usage' style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
+                                    <img src={selectedSkill.meme} />
+                                </div>
+                            )}
+                        </div>
+                    </div>
+                </div>
+
+                <div className={`no-select  ${displayType ? 'light' : 'dark'} ${selectedSkill ? 'hidden' : ''}`} >
+                    <p>Click</p>
+                    <p>A</p>
+                    <p>Skill</p>
+                    <p>to</p>
+                    <p>Learn</p>
+                    <p>More!</p>
+                </div>
+
+                <div className='skill-list' style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
+                    <div className={`skill-header ${displayType ? 'light' : 'dark'}`}>
+                        <h4>Current Tech Stack and Resources</h4>
+                    </div>
+                    <div className='skill-icons'>
+                        {skillsChunk2.map(skill => (
+                            <div
+                                key={skill.name}
+                                className={`skill-item ${displayType ? 'light' : 'dark'}`}
+                                style={{ boxShadow: selectedSkill?.name === skill.name ? `0 0 1rem ${pokeTheme_TWO}` : '' }}
+                                onClick={() => handleSkillClick(skill)}
+                            >
+                                <img
+                                    src={skill.imgSrc}
+                                    alt={skill.name}
+                                    title={skill.name}
+                                    className="skill-icon"
+                                />
+                                <p className={`skill-name ${displayType ? 'light' : 'dark'}`}>{skill.name}</p>
+                            </div>
+                        ))}
                     </div>
                 </div>
             </div>
-
-            <div className={`no-select  ${displayType ? 'light' : 'dark'} ${selectedSkill ? 'hidden' : ''}`} >
-                <p>Click</p>
-                <p>A</p>
-                <p>Skill</p>
-                <p>to</p>
-                <p>Learn</p>
-                <p>More!</p>
-            </div>
-
-            <div className='skill-list' style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}>
-                <div className={`skill-header ${displayType ? 'light' : 'dark'}`}>
-                    <h4>Current Tech Stack and Resources</h4>
-                </div>
-                <div className='skill-icons'>
-                    {skillsChunk2.map(skill => (
-                        <div
-                            key={skill.name}
-                            className={`skill-item ${displayType ? 'light' : 'dark'}`}
-                            style={{ boxShadow: selectedSkill?.name === skill.name ? `0 0 1rem ${pokeTheme_TWO}` : '' }}
-                            onClick={() => handleSkillClick(skill)}
-                        >
-                            <img
-                                src={skill.imgSrc}
-                                alt={skill.name}
-                                title={skill.name}
-                                className="skill-icon"
-                            />
-                            <p className={`skill-name ${displayType ? 'light' : 'dark'}`}>{skill.name}</p>
-                        </div>
-                    ))}
-                </div>
-            </div>
-        </div>
+        </div >
     );
 };
