+++ src/pages/Socials/Socials.jsx
@@ -40,41 +40,43 @@ const Socials = () => {
     };
 
     return (
-        <div className={`socials-main-box`}>
-            <div className='socials-list'>
-                {selectedTab === 3 && (
-                    SOCIALMEDIA.map((item, index) => (
-                        <motion.div
-                            className={`social-media ${displayType ? 'light' : 'dark'}`}
-                            style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}
-                            key={index}
-                            custom={index}
-                            initial="hidden"
-                            animate="visible"
-                            variants={socialMediaVariants}
-                        >
-                            <div className='pic-container-flex-end'>
-                                <div className='sm-pic-container'>
-                                    <img
-                                        src={hoveredImage === index ? item.onHov : item.src}
-                                        onMouseEnter={() => setHoveredImage(index)}
-                                        onMouseLeave={() => setHoveredImage(null)}
-                                        onClick={() => handleImageClick(item.href)}
-                                        style={{ cursor: 'pointer' }}
-                                    />
+        <div className="content-item socials">
+            <div className={`socials-main-box`}>
+                <div className='socials-list'>
+                    {selectedTab === 3 && (
+                        SOCIALMEDIA.map((item, index) => (
+                            <motion.div
+                                className={`social-media ${displayType ? 'light' : 'dark'}`}
+                                style={{ boxShadow: `0 0 0.5em ${pokeTheme_ONE}` }}
+                                key={index}
+                                custom={index}
+                                initial="hidden"
+                                animate="visible"
+                                variants={socialMediaVariants}
+                            >
+                                <div className='pic-container-flex-end'>
+                                    <div className='sm-pic-container'>
+                                        <img
+                                            src={hoveredImage === index ? item.onHov : item.src}
+                                            onMouseEnter={() => setHoveredImage(index)}
+                                            onMouseLeave={() => setHoveredImage(null)}
+                                            onClick={() => handleImageClick(item.href)}
+                                            style={{ cursor: 'pointer' }}
+                                        />
+                                    </div>
                                 </div>
-                            </div>
-                            <div className={`sm-info ${displayType ? 'light' : 'dark'}`}>
-                                <div className='sm-info-li'>
-                                    <a href={item.href} target='_blank' rel="noopener noreferrer" style={{ color: pokeTheme_TWO }}>
-                                        {item.label}
-                                    </a>
-                                    <p>{item.description}</p>
+                                <div className={`sm-info ${displayType ? 'light' : 'dark'}`}>
+                                    <div className='sm-info-li'>
+                                        <a href={item.href} target='_blank' rel="noopener noreferrer" style={{ color: pokeTheme_TWO }}>
+                                            {item.label}
+                                        </a>
+                                        <p>{item.description}</p>
+                                    </div>
                                 </div>
-                            </div>
-                        </motion.div>
-                    ))
-                )}
+                            </motion.div>
+                        ))
+                    )}
+                </div>
             </div>
         </div>
     );
