--- src/utils/PokeColors.jsx
@@ -1,20 +1,0 @@
-export const POKE_COLORS = {
-    normal: { base: '#8D8D6E', complementary: '#6E6E8D' }, // Muted Olive + Muted Purple
-    fire: { base: '#D9531E', complementary: '#1E7AD9' }, // Darker Orange + Deep Blue
-    water: { base: '#2C6AC3', complementary: '#C32C6A' }, // Medium Blue + Reddish Pink
-    electric: { base: '#C2A500', complementary: '#0058C2' }, // Rich Yellow + Deep Blue
-    grass: { base: '#528B3E', complementary: '#3E528B' }, // Darker Green + Deep Blue
-    ice: { base: '#5B9EA6', complementary: '#A65B69' }, // Cool Blue-Gray + Warm Red-Gray
-    fighting: { base: '#A73220', complementary: '#2042A7' }, // Deep Red + Deep Blue
-    poison: { base: '#834285', complementary: '#428548' }, // Muted Purple + Muted Green
-    ground: { base: '#A57E4A', complementary: '#4A78A5' }, // Warm Brown + Muted Blue
-    flying: { base: '#7C6DB0', complementary: '#B07C6D' }, // Muted Purple + Muted Warm Coral
-    psychic: { base: '#C43868', complementary: '#38C4A5' }, // Deep Pink + Muted Teal
-    bug: { base: '#8A9A1E', complementary: '#1E8A9A' }, // Olive Green + Teal
-    rock: { base: '#8A7B49', complementary: '#49748A' }, // Earthy Brown + Muted Blue-Gray
-    ghost: { base: '#584D82', complementary: '#826E4D' }, // Dark Purple + Muted Brown
-    dragon: { base: '#5A38A7', complementary: '#38A76E' }, // Deep Purple + Muted Green
-    dark: { base: '#534741', complementary: '#415354' }, // Dark Brown + Muted Teal-Gray
-    steel: { base: '#88889B', complementary: '#9B8988' }, // Cool Gray + Warm Gray
-    fairy: { base: '#B36787', complementary: '#67B399' }, // Muted Pink + Soft Green
-};
